Move mixins in hooks.py to mixins.py
