+++ wagtail/admin/views/generic/__init__.py
@@ -1,4 +1,8 @@
-from .hooks import BeforeAfterHookMixin, HookResponseMixin  # noqa
-from .mixins import LocaleMixin  # noqa
+from .mixins import (
+    BeforeAfterHookMixin,
+    HookResponseMixin,
+    LocaleMixin,
+    PanelMixin,
+)  # noqa
 from .models import CreateView, DeleteView, EditView, IndexView  # noqa
 from .permissions import PermissionCheckedMixin  # noqa
