+++ wagtail/admin/views/generic/mixins.py
@@ -2,7 +2,73 @@ from django.conf import settings
 from django.forms import Media
 from django.shortcuts import get_object_or_404
 
+from wagtail import hooks
 from wagtail.models import Locale, TranslatableMixin
+
+
+class HookResponseMixin:
+    """
+    A mixin for class-based views to run hooks by `hook_name`.
+    """
+
+    def run_hook(self, hook_name, *args, **kwargs):
+        """
+        Run the named hook, passing args and kwargs to each function registered under that hook name.
+        If any return an HttpResponse, stop processing and return that response
+        """
+        for fn in hooks.get_hooks(hook_name):
+            result = fn(*args, **kwargs)
+            if hasattr(result, "status_code"):
+                return result
+        return None
+
+
+class BeforeAfterHookMixin(HookResponseMixin):
+    """
+    A mixin for class-based views to support hooks like `before_edit_page` and
+    `after_edit_page`, which are triggered during execution of some operation and
+    can return a response to halt that operation and/or change the view response.
+    """
+
+    def run_before_hook(self):
+        """
+        Define how to run the hooks before the operation is executed.
+        The `self.run_hook(hook_name, *args, **kwargs)` from HookResponseMixin
+        can be utilised to call the hooks.
+
+        If this method returns a response, the operation will be aborted and the
+        hook response will be returned as the view response, skipping the default
+        response.
+        """
+        return None
+
+    def run_after_hook(self):
+        """
+        Define how to run the hooks after the operation is executed.
+        The `self.run_hook(hook_name, *args, **kwargs)` from HookResponseMixin
+        can be utilised to call the hooks.
+
+        If this method returns a response, it will be returned as the view
+        response immediately after the operation finishes, skipping the default
+        response.
+        """
+        return None
+
+    def dispatch(self, *args, **kwargs):
+        hooks_result = self.run_before_hook()
+        if hooks_result is not None:
+            return hooks_result
+
+        return super().dispatch(*args, **kwargs)
+
+    def form_valid(self, form):
+        response = super().form_valid(form)
+
+        hooks_result = self.run_after_hook()
+        if hooks_result is not None:
+            return hooks_result
+
+        return response
 
 
 class LocaleMixin:
